@@ -29,7 +29,7 @@
     }
   },
   "require": {
-    "php": ">=7.2.5",
+    "php": ">=7.4",
     "ext-json": "*",
     "eftec/clione": "^1.31",
     "ext-pdo": "*",
@@ -44,7 +44,7 @@
     "eftec/validationone": "For keeping and storing the messages"
   },
   "require-dev": {
-    "phpunit/phpunit": "^8.5.13"
+    "phpunit/phpunit": "^8.5.36"
   }
 }
 
